+++ node/src/types/kernel/api/notebook/openNotebook/index.d.ts
@@ -17,9 +17,7 @@
 
 
 /* directories */
-export * as closeNotebook from "./closeNotebook";
-export * as getNotebookConf from "./getNotebookConf";
-export * as lsNotebooks from "./lsNotebooks";
-export * as openNotebook from "./openNotebook";
 
 /* flles */
+export * from "./payload";
+export * from "./response";