+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - 添加 API `/api/notebook/lsNotebooks` | Add API `/api/notebook/lsNotebooks`
 - 添加 API `/api/notebook/getNotebookConf` | Add API `/api/notebook/getNotebookConf`
 - 添加 API `/api/notebook/closeNotebook` | Add API `/api/notebook/closeNotebook`
+- 添加 API `/api/notebook/openNotebook` | Add API `/api/notebook/openNotebook`
 
 ## v0.0.4 / 2023-06-29